Data loading

Data loading

Usage

kobo_data(datapath)

Arguments

datapath

path to the file with the data format as extracted from kobo with dot as group separator and xml header

Value

A "datalist" S3 class object (list) formatted to the specifications of "kobocruncher".

Examples

datalist <- kobo_data(datapath = system.file("data.xlsx", package = "kobocruncher") )
# MainFrame
datalist[["main"]]
#> # A tibble: 5 × 26
#>   start               end                 location       profi…¹ profi…² profi…³
#>   <dttm>              <dttm>              <chr>          <chr>   <chr>   <chr>  
#> 1 2022-10-27 08:57:57 2022-11-10 08:49:09 private_facil… VEN     Consul… accomo…
#> 2 2022-10-24 11:36:52 2022-10-24 15:46:00 home           HND     farmer  accomo…
#> 3 2022-10-26 16:19:35 2022-10-26 18:15:07 subcenter      SLV     vendor  educat…
#> 4 2022-10-26 16:19:35 2022-10-26 17:51:14 subcenter      SLV     teacher safety…
#> 5 2022-10-26 14:02:06 2022-10-26 16:10:25 home           COL     farmer  employ…
#> # … with 20 more variables: profile.reason.accomodation <dbl>,
#> #   profile.reason.employment <dbl>, profile.reason.education <dbl>,
#> #   profile.reason.community <dbl>, profile.reason.safety <dbl>,
#> #   profile.reason.movement <dbl>, profile.reason.reunification <dbl>,
#> #   profile.reason.no_answer <dbl>, profile.reason.other <dbl>,
#> #   profile.HHSize <dbl>, `_id` <dbl>, `_uuid` <chr>, `_submission_time` <dbl>,
#> #   `_validation_status` <lgl>, `_notes` <lgl>, `_status` <chr>, …
# Second Frame - based on presence of repeat within the form, aka nested or
# hierarchical data structure, etc... 
datalist[["members"]]
#> # A tibble: 12 × 14
#>    members.age members…¹ _inde…² _pare…³ _pare…⁴ _subm…⁵ _subm…⁶ _subm…⁷ _subm…⁸
#>          <dbl> <chr>       <dbl> <chr>     <dbl>   <dbl> <chr>     <dbl> <lgl>  
#>  1          56 male            1 Sample…       1  2.08e7 48cc75…  44686. NA     
#>  2           2 male            2 Sample…       2  2.08e7 f1c3d3…  44686. NA     
#>  3           3 male            3 Sample…       2  2.08e7 f1c3d3…  44686. NA     
#>  4          10 female          4 Sample…       2  2.08e7 f1c3d3…  44686. NA     
#>  5          45 male            5 Sample…       2  2.08e7 f1c3d3…  44686. NA     
#>  6          35 female          6 Sample…       2  2.08e7 f1c3d3…  44686. NA     
#>  7           4 male            7 Sample…       3  2.08e7 1bac72…  44686. NA     
#>  8          34 female          8 Sample…       3  2.08e7 1bac72…  44686. NA     
#>  9          34 male            9 Sample…       3  2.08e7 1bac72…  44686. NA     
#> 10          51 female         10 Sample…       4  2.08e7 aa9f07…  44686. NA     
#> 11          21 male           11 Sample…       5  2.08e7 e83161…  44686. NA     
#> 12          25 female         12 Sample…       5  2.08e7 e83161…  44686. NA     
#> # … with 5 more variables: `_submission__notes` <lgl>,
#> #   `_submission__status` <chr>, `_submission__submitted_by` <lgl>,
#> #   `_submission__tags` <lgl>, parent_index <dbl>, and abbreviated variable
#> #   names ¹​members.sex, ²​`_index`, ³​`_parent_table_name`, ⁴​`_parent_index`,
#> #   ⁵​`_submission__id`, ⁶​`_submission__uuid`, ⁷​`_submission__submission_time`,
#> #   ⁸​`_submission__validation_status`