Prepare XLSform

Prepare XLSform by adding instructions for the analysis plan and checking that structure and settings are correct.

Once those elements are set up, they will be automatically considered during the automatic crunching phase. An additional worksheet is also created to document the information required for registration on UNHCR CKAN instance http://ridl.unhcr.org

Configuration of how questions are grouped together in the report:

chapter: by default the crunching report is presented according to the group. Once set, this will replace the original grouping. Only variable defined within a chapter will be displayed in the crunching report.

Configuration for data manipulation:

clean: define what variable shall be re-categorized during cleaning - a local copy of all levels will be locally saved in order to do the mapping in excel. When the mapping is available, it will be automatically applied to the data. Can be useful to reduce the number of categories.

anonymise: define what variables to consider to statistical disclosure risk measurement and subsequent data treatment

Configuration of specific charts, visualization and analysis:

disaggregation: define variable to use for visual crosstabulation - functions with "_cross"

correlate: define the variable to use to explore statistical association - works under certan restrictions (i.e. between 2 categorical variables only): kobo_correlate

cluster: define variable to generate an unsupervised classification (i.e. hierarchical clustering based on multiple correspondance analysis) kobo_cluster

predict: define variable to use to generate predictive model, ie.e the target variable and the predictors. kobo_predict

mappoint, mappoly: define the variable to use to generate maps - kobo_map

In case if those fields do not yet exist, the function will create dummy column for each one. Also, coloring all rows that have type equal to "begin group", "end group", "begin repeat" or "end repeat" for better legibility

Usage

kobo_prepare_form(xlsformpath, xlsformpathout, language = "::english (en)")

Arguments

xlsformpath

The full path and filename of the xlsform to be accessed (has to be xlsx file)

xlsformpathout

The full path and filename of the xlsform to be accessed (has to be xlsx file)

language

Optional if the form used multiple languages, indicate the language to use to prepare the analysis plan - check first in your original file -default is ::english (en)