Initial Template for Automatic Data Exploration

This template is designed for initial data crunching - The first RMD template gives an output in HTML for easy navigation - the left menu provides smooth transition. It includes a function to automatically run throughout all the survey content. During this stage, data cleaning and new variable creation can be performed through iterations This report also includes each plot syntax so that they can be easily pasted for the second report

Usage

template_1_exploration(
  datafolder = "data-raw",
  ridl = "ridlproject",
  data = "data.xlsx",
  form = "form.xlsx",
  datasource = "Study name reference",
  publish = "no",
  republish = "no",
  visibility = "public",
  stage = "exploration_initial",
  language = "",
  folder = "Report"
)

Arguments

datafolder

"data-raw" This is the default folder where to put you data in

ridl

"ridlproject" Name of the ridl project where you data is documented and archived

data

"data.xlsx" Name of the data file

form

"form.xlsx" Name of the xlsform -

datasource

"Study name reference" ## String used in caption for all your charts

publish

"no" Put to "yes" in order to add your report, source and analysis plan as ressource within the same ridl c

republish

"no"

visibility

"public"

stage

"exploration_initial" You may change this to exploration_advanced if you configuring many

language

"" Check what you have in your xlsform - ::english (en) -or ::french (fr) or ::spanish (es)

folder

folder within your project where to put the generated report. Folder will be created if it does not exist

Value

nothing the file for the report is generated

Examples

# template_1_exploration(datafolder= "data-raw",
#                                    ridl =  "ridlproject",
#                                     data = "data.xlsx" ,
#                                     form =  "form.xlsx",
#                                     datasource = "Study name reference",
#                                     publish =  "no", 
#                                     republish = "no",
#                                     visibility = "public",
#                                     stage = "exploration_initial",
#                                    language = "",
#                                    folder = "Report")